+++ qradar/release_notes.md
@@ -1,4 +1,13 @@
-#### What's Fixed
-  
-- Updated jinja .timestamp to .int_timestamp in the data ingestion playbooks.
-- Added the "PATCH" option to the "Request Method" parameter of the "Invoke QRadar REST API" operation.
+#### What's new in 1.6.0
+
+- New actions:
+    - Get assets properties: Get a list of available asset property types that can be used with assets update operation
+    - Get assets: Get a list of available assets
+    - Update asset: Update one or multiple attributes of a specific asset defined by Asset ID
+    - Get cases: Get a list of Qradar cases
+    - Create case: Create a new case on Qradar
+    - Get reference tables: Get a list of Reference Tables
+    - Get table elements: Get records details of a reference table defined by name
+    - Add table element: Add or update an element in a reference table
+    - Delete table element: Delete element from a reference table
+    - Delete reference table: Delete a Reference Table or Purge its content